created a rating util file to store function and render star component accordingly to average rating
